ajout saisisseur existant

+++ app/Http/Controllers/MutantController.php
@@ -2,13 +2,16 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Region;
 use App\Models\Commune;
+use App\Models\District;
+//use App\Models\Proprietaire;
+use App\Models\Province;
 use App\Models\Fokontany;
 use App\Models\Nationalite;
-//use App\Models\Proprietaire;
+use App\Models\Proprietaire;
 use Illuminate\Http\Request;
 use App\Models\Etablissement;
-use App\Models\Proprietaire;
 use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
@@ -64,25 +67,24 @@ class MutantController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit($id)  
     {
         $etablissement = Etablissement::with('proprietaires')->find($id);
-        $proprietaire = $etablissement->proprietaires->first();
-
-        $communes = DB::table('communes')
-        ->select('commune', 'id')
-        ->where('region', '=', Auth()->user()->region_user)
-        ->groupBy('commune')
-        ->get();
-        
-        $fokontany = Fokontany::all();
+       // dd($etablissement);
+         $provinces =Province::all();
+         $regions = Region::all();
+         $districts = District::all();
+         $communes = Commune::all();
+         $fokontany = Fokontany::all();
         $nationalites = Nationalite::all();
-
+        
         return view('saisisseur.mutation.edit')
-            ->with('fokontanys', $fokontany)
+            ->with('provinces', $provinces)
+            ->with('regions', $regions)
+            ->with('districts', $districts)
             ->with('communes', $communes)
+            ->with('fokontany', $fokontany)
             ->with('nationalites', $nationalites)
-            ->with('proprietaire', $proprietaire)
             ->with('etablissement', $etablissement);
     }
        
@@ -95,54 +97,52 @@ class MutantController extends Controller
      */
     public function mutation_ft(Request $request, $id, $prop_id)
     {
-        //dd($request);
         $validator = Validator::make($request->all(), [
-
             'cin' => 'required',
             'nom' => 'required',
             'adresse' => 'required',
             'lien' => 'required',
-            'num_tel' => 'required',
-            'fokotany_proprietaire' => 'required|exists:fokontanies,id',
-            'commune_proprio' => 'required|exists:communes,id',
-            'nationalite_proprietaire' => 'required|exists:nationalites,id',
+            'num_tel' => 'required', 
+            'province' => 'required|exists:provinces,id',
+            'region' => 'required|exists:regions,id',
+            'district' => 'required|exists:districts,id',
+            'commune' => 'required|exists:communes,id',
+            'fokontany' => 'required|exists:fokontanies,id',
+            'nationalite_id' => 'required|exists:nationalites,id',
         ]);
+    
         if ($validator->fails()) {
             return redirect()->back()->withErrors($validator)->withInput();
         }
-
-        $etablissement = Etablissement::findOrFail($id);
+    
+        $etablissement = Etablissement::with('proprietaires')->findOrFail($id);
         $ancien_proprietaire = $etablissement->proprietaires->first();
-        // if (intval($ancien_proprietaire->lien))
-        // {
-        $ancien_proprietaire->lien = $ancien_proprietaire->lien - 1;
-        $etablissement->proprietaires()->detach($ancien_proprietaire->id);
-        // }
-        // else
-        // {
-        //     $ancien_proprietaire->delete();
-        // }
-        $ancien_proprietaire->save();
-        
-        //proprietaire
+        if ($ancien_proprietaire) {
+            $ancien_proprietaire->lien = $ancien_proprietaire->lien - 1;
+            $etablissement->proprietaires()->detach($ancien_proprietaire->id);
+            $ancien_proprietaire->save();
+        }
+    
         $proprietaire = new Proprietaire();
         $proprietaire->cin = $request->input('cin');
         $proprietaire->nom = $request->input('nom');
         $proprietaire->adresse = $request->input('adresse');
-        $proprietaire->nationalite_id = $request->input('nationalite_proprietaire');
-        $proprietaire->fokontany_id = $request->input('fokotany_proprietaire');
-        $proprietaire->commune_id = $request->input('commune_proprio');
+        $proprietaire->nationalite_id = $request->input('nationalite_id');
+        $proprietaire->province_id = $request->input('province');
+        $proprietaire->region_id = $request->input('region');
+        $proprietaire->district_id = $request->input('district');
+        $proprietaire->fokontany_id = $request->input('fokontany');
+        $proprietaire->commune_id = $request->input('commune');
         $proprietaire->lien = 0;
         $proprietaire->email = $request->input('email');
         $proprietaire->num_tel = $request->input('num_tel');
-
-        //etablissement
+        $proprietaire->save();
+    
         $etablissement->status = "En attente";
         $etablissement->type = "U";
         $etablissement->save();
-        $proprietaire->save();
         $etablissement->proprietaires()->attach($proprietaire->id);
-
+        dd($etablissement);
         return redirect()->route('mutation.index')->with('success', 'Données envoyées avec succès !!!');
